+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.3.0
 - GetSpritesPalette functionality added.
+- Corrected break reason.
 
 ## 0.2.0
 - Changed to new CSpect API 2.12.20.